+++ src/main/java/pl/weather/model/Daily.java
@@ -37,15 +37,15 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Daily {
 
     @JsonProperty("dt")
-    private Integer dt;
+    private Double dt;
     @JsonProperty("sunrise")
-    private Integer sunrise;
+    private Double sunrise;
     @JsonProperty("sunset")
-    private Integer sunset;
+    private Double sunset;
     @JsonProperty("moonrise")
-    private Integer moonrise;
+    private Double moonrise;
     @JsonProperty("moonset")
-    private Integer moonset;
+    private Double moonset;
     @JsonProperty("moon_phase")
     private Double moonPhase;
     @JsonProperty("temp")
@@ -53,21 +53,21 @@ public class Daily {
     @JsonProperty("feels_like")
     private FeelsLike feelsLike;
     @JsonProperty("pressure")
-    private Integer pressure;
+    private Double pressure;
     @JsonProperty("humidity")
-    private Integer humidity;
+    private Double humidity;
     @JsonProperty("dew_point")
     private Double dewPoint;
     @JsonProperty("wind_speed")
     private Double windSpeed;
     @JsonProperty("wind_deg")
-    private Integer windDeg;
+    private Double windDeg;
     @JsonProperty("wind_gust")
     private Double windGust;
     @JsonProperty("weather")
     private List<Weather__1> weather = null;
     @JsonProperty("clouds")
-    private Integer clouds;
+    private Double clouds;
     @JsonProperty("pop")
     private Double pop;
     @JsonProperty("rain")
@@ -78,52 +78,52 @@ public class Daily {
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("dt")
-    public Integer getDt() {
+    public Double getDt() {
         return dt;
     }
 
     @JsonProperty("dt")
-    public void setDt(Integer dt) {
+    public void setDt(Double dt) {
         this.dt = dt;
     }
 
     @JsonProperty("sunrise")
-    public Integer getSunrise() {
+    public Double getSunrise() {
         return sunrise;
     }
 
     @JsonProperty("sunrise")
-    public void setSunrise(Integer sunrise) {
+    public void setSunrise(Double sunrise) {
         this.sunrise = sunrise;
     }
 
     @JsonProperty("sunset")
-    public Integer getSunset() {
+    public Double getSunset() {
         return sunset;
     }
 
     @JsonProperty("sunset")
-    public void setSunset(Integer sunset) {
+    public void setSunset(Double sunset) {
         this.sunset = sunset;
     }
 
     @JsonProperty("moonrise")
-    public Integer getMoonrise() {
+    public Double getMoonrise() {
         return moonrise;
     }
 
     @JsonProperty("moonrise")
-    public void setMoonrise(Integer moonrise) {
+    public void setMoonrise(Double moonrise) {
         this.moonrise = moonrise;
     }
 
     @JsonProperty("moonset")
-    public Integer getMoonset() {
+    public Double getMoonset() {
         return moonset;
     }
 
     @JsonProperty("moonset")
-    public void setMoonset(Integer moonset) {
+    public void setMoonset(Double moonset) {
         this.moonset = moonset;
     }
 
@@ -158,22 +158,22 @@ public class Daily {
     }
 
     @JsonProperty("pressure")
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
     @JsonProperty("pressure")
-    public void setPressure(Integer pressure) {
+    public void setPressure(Double pressure) {
         this.pressure = pressure;
     }
 
     @JsonProperty("humidity")
-    public Integer getHumidity() {
+    public Double getHumidity() {
         return humidity;
     }
 
     @JsonProperty("humidity")
-    public void setHumidity(Integer humidity) {
+    public void setHumidity(Double humidity) {
         this.humidity = humidity;
     }
 
@@ -198,12 +198,12 @@ public class Daily {
     }
 
     @JsonProperty("wind_deg")
-    public Integer getWindDeg() {
+    public Double getWindDeg() {
         return windDeg;
     }
 
     @JsonProperty("wind_deg")
-    public void setWindDeg(Integer windDeg) {
+    public void setWindDeg(Double windDeg) {
         this.windDeg = windDeg;
     }
 
@@ -228,12 +228,12 @@ public class Daily {
     }
 
     @JsonProperty("clouds")
-    public Integer getClouds() {
+    public Double getClouds() {
         return clouds;
     }
 
     @JsonProperty("clouds")
-    public void setClouds(Integer clouds) {
+    public void setClouds(Double clouds) {
         this.clouds = clouds;
     }
